% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ciu.image.R
\name{ciu.image.new}
\alias{ciu.image.new}
\title{ciu.image.new}
\usage{
ciu.image.new(model, predict.function = NULL)
}
\arguments{
\item{model}{blabla}

\item{predict.function}{blabla}
}
\value{
blabla
}
\description{
ciu.image.new
}
