% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ciu.image.result.R
\name{ciu.image.result.new}
\alias{ciu.image.result.new}
\title{ciu.image.result.new}
\usage{
ciu.image.result.new(ci, cu, cmin, cmax, outval, out.names)
}
\arguments{
\item{ci}{blabla}

\item{cu}{blabla}

\item{cmin}{blabla}

\item{cmax}{blabla}

\item{outval}{blabla}

\item{out.names}{blabla}
}
\value{
blabla
}
\description{
ciu.image.result.new
}
