% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ciu.image.result.R
\name{ciu.image.result.new}
\alias{ciu.image.result.new}
\title{Create \code{ciu.image.result} object.}
\usage{
ciu.image.result.new(ci, cu, cmin, cmax, outval, out.names)
}
\arguments{
\item{ci}{CI values.}

\item{cu}{CU values.}

\item{cmin}{Cmin values.}

\item{cmax}{Cmax values.}

\item{outval}{Output values.}

\item{out.names}{Output names.}
}
\value{
blabla
}
\description{
A \code{ciu.image.result} object is a \link{data.frame} with columns that correspond
to the passed parameters. There is one value/row per output.
}
